# Conflicts:
#	src/main/java/net/minecraft/server/BlockSapling.java
#	src/main/java/net/minecraft/server/EntitySkeleton.java
#	src/main/java/net/minecraft/server/MinecraftServer.java
#	src/main/java/net/minecraft/server/NetHandler.java
#	src/main/java/net/minecraft/server/NetServerHandler.java
#	src/main/java/net/minecraft/server/RegionFile.java
#	src/main/java/net/minecraft/server/RegionFileWAL.java
#	src/main/java/net/minecraft/server/ServerProfiler.java
#	src/main/java/net/minecraft/server/VoiceChatUDPServer.java
#	src/main/java/net/minecraft/server/network/ModProtocol.java

+++ src/main/java/net/minecraft/server/WorldLoaderServer.java
@@ -19,7 +19,21 @@ public class WorldLoaderServer extends WorldLoader {
     public boolean isConvertable(String s) {
         WorldData worlddata = this.b(s);
 
-        return worlddata != null && worlddata.i() == 0;
+        return worlddata != null && worlddata.i() == WorldSaveVersions.LEGACY_PRE_MCREGION;
+    }
+
+    public boolean hasLegacyChunkData(String s) {
+        File file1 = new File(this.a, s);
+        if (!file1.exists() || !file1.isDirectory()) {
+            return false;
+        }
+
+        if (this.containsLegacyChunkFiles(file1)) {
+            return true;
+        }
+
+        File file2 = new File(file1, "DIM-1");
+        return this.containsLegacyChunkFiles(file2);
     }
 
     public boolean convert(String s, IProgressUpdate iprogressupdate) {
@@ -30,6 +44,9 @@ public class WorldLoaderServer extends WorldLoader {
         ArrayList arraylist3 = new ArrayList();
         File file1 = new File(this.a, s);
         File file2 = new File(file1, "DIM-1");
+        if (!file1.exists() || !file1.isDirectory()) {
+            return false;
+        }
 
         System.out.println("Scanning folders...");
         this.a(file1, arraylist, arraylist1);
@@ -41,10 +58,15 @@ public class WorldLoaderServer extends WorldLoader {
 
         System.out.println("Total conversion count is " + i);
         this.a(file1, arraylist, 0, i, iprogressupdate);
-        this.a(file2, arraylist2, arraylist.size(), i, iprogressupdate);
+        if (file2.exists()) {
+            this.a(file2, arraylist2, arraylist.size(), i, iprogressupdate);
+        }
         WorldData worlddata = this.b(s);
+        if (worlddata == null) {
+            throw new RuntimeException("Failed to load world metadata after legacy chunk conversion for '" + s + "'");
+        }
 
-        worlddata.a(19132);
+        worlddata.a(WorldSaveVersions.MCREGION_1);
         IDataManager idatamanager = this.a(s, false);
 
         idatamanager.a(worlddata);
@@ -52,14 +74,66 @@ public class WorldLoaderServer extends WorldLoader {
         if (file2.exists()) {
             this.a(arraylist3, arraylist.size() + arraylist2.size() + arraylist1.size(), i, iprogressupdate);
         }
+        try {
+            // Keep conversion parity with client RegionCore upgrade logic so chest inventories/facing
+            // and other legacy item-bearing NBT are rewritten in one deterministic pass.
+            RegionCoreWorldUpgrader.upgradeWorldToRegionCore(file1, MinecraftServer.log);
+        } catch (RuntimeException runtimeexception) {
+            throw new RuntimeException("Failed to run RegionCore post-conversion upgrade for '" + s + "'", runtimeexception);
+        }
+
+        iprogressupdate.a(100);
 
         return true;
     }
 
-    private void a(File file1, ArrayList arraylist, ArrayList arraylist1) {
+    private boolean containsLegacyChunkFiles(File file1) {
+        if (file1 == null || !file1.exists() || !file1.isDirectory()) {
+            return false;
+        }
+
         ChunkFileFilter chunkfilefilter = new ChunkFileFilter((EmptyClass2) null);
         ChunkFilenameFilter chunkfilenamefilter = new ChunkFilenameFilter((EmptyClass2) null);
         File[] afile = file1.listFiles(chunkfilefilter);
+        if (afile == null) {
+            return false;
+        }
+
+        for (int i = 0; i < afile.length; ++i) {
+            File file2 = afile[i];
+            if (file2 == null) {
+                continue;
+            }
+            File[] afile2 = file2.listFiles(chunkfilefilter);
+            if (afile2 == null) {
+                continue;
+            }
+
+            for (int j = 0; j < afile2.length; ++j) {
+                File file3 = afile2[j];
+                if (file3 == null) {
+                    continue;
+                }
+                File[] afile3 = file3.listFiles(chunkfilenamefilter);
+                if (afile3 != null && afile3.length > 0) {
+                    return true;
+                }
+            }
+        }
+
+        return false;
+    }
+
+    private void a(File file1, ArrayList arraylist, ArrayList arraylist1) {
+        if (file1 == null || !file1.exists() || !file1.isDirectory()) {
+            return;
+        }
+        ChunkFileFilter chunkfilefilter = new ChunkFileFilter((EmptyClass2) null);
+        ChunkFilenameFilter chunkfilenamefilter = new ChunkFilenameFilter((EmptyClass2) null);
+        File[] afile = file1.listFiles(chunkfilefilter);
+        if (afile == null) {
+            return;
+        }
         File[] afile1 = afile;
         int i = afile.length;
 
@@ -68,12 +142,18 @@ public class WorldLoaderServer extends WorldLoader {
 
             arraylist1.add(file2);
             File[] afile2 = file2.listFiles(chunkfilefilter);
+            if (afile2 == null) {
+                continue;
+            }
             File[] afile3 = afile2;
             int k = afile2.length;
 
             for (int l = 0; l < k; ++l) {
                 File file3 = afile3[l];
                 File[] afile4 = file3.listFiles(chunkfilenamefilter);
+                if (afile4 == null) {
+                    continue;
+                }
                 File[] afile5 = afile4;
                 int i1 = afile4.length;
 
@@ -91,38 +171,52 @@ public class WorldLoaderServer extends WorldLoader {
         byte[] abyte = new byte[4096];
         Iterator iterator = arraylist.iterator();
 
-        while (iterator.hasNext()) {
-            ChunkFile chunkfile = (ChunkFile) iterator.next();
-            int k = chunkfile.b();
-            int l = chunkfile.c();
-            RegionFile regionfile = RegionFileCache.a(file1, k, l);
+        try {
+            while (iterator.hasNext()) {
+                ChunkFile chunkfile = (ChunkFile) iterator.next();
+                int k = chunkfile.b();
+                int l = chunkfile.c();
+                RegionFile regionfile = RegionFileCache.a(file1, k, l);
 
-            if (!regionfile.c(k & 31, l & 31)) {
-                try {
-                    DataInputStream datainputstream = new DataInputStream(new GZIPInputStream(new FileInputStream(chunkfile.a())));
-                    DataOutputStream dataoutputstream = regionfile.b(k & 31, l & 31);
-                    boolean flag = false;
+                if (!regionfile.c(k & 31, l & 31)) {
+                    DataInputStream datainputstream = null;
+                    DataOutputStream dataoutputstream = null;
 
-                    int i1;
+                    try {
+                        datainputstream = new DataInputStream(new GZIPInputStream(new FileInputStream(chunkfile.a())));
+                        dataoutputstream = regionfile.b(k & 31, l & 31);
 
-                    while ((i1 = datainputstream.read(abyte)) != -1) {
-                        dataoutputstream.write(abyte, 0, i1);
+                        int i1;
+                        while ((i1 = datainputstream.read(abyte)) != -1) {
+                            dataoutputstream.write(abyte, 0, i1);
+                        }
+                    } catch (IOException ioexception) {
+                        throw new RuntimeException("Failed converting legacy chunk '" + chunkfile.a().getAbsolutePath()
+                                + "' into region coordinates (" + k + "," + l + ")", ioexception);
+                    } finally {
+                        if (dataoutputstream != null) {
+                            try {
+                                dataoutputstream.close();
+                            } catch (IOException ignored) {}
+                        }
+                        if (datainputstream != null) {
+                            try {
+                                datainputstream.close();
+                            } catch (IOException ignored) {}
+                        }
                     }
+                }
 
-                    dataoutputstream.close();
-                    datainputstream.close();
-                } catch (IOException ioexception) {
-                    ioexception.printStackTrace();
+                ++i;
+                if (j > 0) {
+                    int j1 = (int) Math.round(100.0D * (double) i / (double) j);
+
+                    iprogressupdate.a(j1);
                 }
             }
-
-            ++i;
-            int j1 = (int) Math.round(100.0D * (double) i / (double) j);
-
-            iprogressupdate.a(j1);
+        } finally {
+            RegionFileCache.a();
         }
-
-        RegionFileCache.a();
     }
 
     private void a(ArrayList arraylist, int i, int j, IProgressUpdate iprogressupdate) {
@@ -130,14 +224,21 @@ public class WorldLoaderServer extends WorldLoader {
 
         while (iterator.hasNext()) {
             File file1 = (File) iterator.next();
+            if (file1 == null) {
+                continue;
+            }
             File[] afile = file1.listFiles();
 
-            a(afile);
+            if (afile != null) {
+                a(afile);
+            }
             file1.delete();
             ++i;
-            int k = (int) Math.round(100.0D * (double) i / (double) j);
+            if (j > 0) {
+                int k = (int) Math.round(100.0D * (double) i / (double) j);
 
-            iprogressupdate.a(k);
+                iprogressupdate.a(k);
+            }
         }
     }
 }
